% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-binomial.R
\name{avoid_table}
\alias{avoid_table}
\title{Avoidance table}
\usage{
avoid_table(B, s)
}
\arguments{
\item{B, s}{Non-negative integers.}
}
\value{
Numeric vector of length \code{B + 1}.
}
\description{
\code{avoid_table(B, s)[k + 1] = choose(B - k, s) / choose(B, s)}, with
value zero whenever \code{B - k < s}.
}
\keyword{internal}
