% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{.lapply_cores}
\alias{.lapply_cores}
\title{Parallel lapply helper}
\usage{
.lapply_cores(X, FUN, n_cores = 1L)
}
\arguments{
\item{X}{List/vector to iterate over.}

\item{FUN}{Function applied to each element.}

\item{n_cores}{Number of workers (forked on unix, PSOCK otherwise).}
}
\value{
A list.
}
\description{
Parallel lapply helper
}
\keyword{internal}
