% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mark_levels}
\alias{.mark_levels}
\title{Mark levels of a point pattern}
\usage{
.mark_levels(pp)
}
\arguments{
\item{pp}{A \code{ppp} object.}
}
\value{
Character vector of levels, or \code{NULL} when unmarked.
}
\description{
Mark levels of a point pattern
}
\keyword{internal}
