% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.subset_mark}
\alias{.subset_mark}
\title{Subset a point pattern to one mark level}
\usage{
.subset_mark(pp, lvl)
}
\arguments{
\item{pp}{A marked \code{ppp}.}

\item{lvl}{Mark level to keep.}
}
\value{
A \code{ppp} containing only points with that mark.
}
\description{
Subset a point pattern to one mark level
}
\keyword{internal}
