% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnG-class.R
\name{nnG-methods}
\alias{nnG-methods}
\alias{as.data.frame.nnG}
\alias{print.nnG}
\alias{summary.nnG}
\alias{plot.nnG}
\title{Methods for nnG objects}
\usage{
\method{as.data.frame}{nnG}(x, row.names = NULL, optional = FALSE, ...)

\method{print}{nnG}(x, ...)

\method{summary}{nnG}(object, ...)

\method{plot}{nnG}(x, y = NULL, ...)
}
\arguments{
\item{x, object}{An \code{nnG} object.}

\item{row.names, optional, ...}{Further arguments (ignored or passed on).}

\item{y}{Ignored; present for compatibility with the plot generic.}
}
\value{
\code{as.data.frame} returns the underlying data frame,
  \code{summary} a list of class \code{summary.nnG}, and
  \code{print}, \code{plot} return their input invisibly.
}
\description{
Methods for nnG objects
}
